+++ libs/library/src/lib/helpers/index.ts
@@ -1,3 +1,4 @@
 export { DateHelpers } from './date-helpers';
+export { DisplayControls } from './display-controls';
 export { GuidHelper } from './guid-helper';
 export { RouterFormatter } from './route-formatter';